Rename `auto_types` feature flag
Rename `auto_types` feature flag to `auto_into_responses` since it
actually is only used for automatic IntoResponses definition. This
feature is higly experiemental and should be used with cauntion. It is
more likely that the final implementation will look different thus
changes are to be expected.

Simultaneusly allow automatic resolvation for request bodies without
extra feature flag. Supported only with `actix_extras` feature by now.
