+++ scripts/test.sh
@@ -11,14 +11,14 @@ if [[ "$crate" == "utoipa" ]]; then
 elif [[ "$crate" == "utoipa-gen" ]]; then
     cargo test -p utoipa-gen --features utoipa/actix_extras,chrono,decimal,utoipa/uuid,utoipa/time,time,utoipa/repr
 
-    cargo test -p utoipa-gen --test path_derive_auto_types --features auto_types
+    cargo test -p utoipa-gen --test path_derive_auto_into_responses --features auto_into_responses
     cargo test -p utoipa-gen --test path_derive_actix --test path_parameter_derive_actix --features actix_extras
-    cargo test -p utoipa-gen --test path_derive_auto_types_actix --features actix_extras,auto_types
+    cargo test -p utoipa-gen --test path_derive_auto_into_responses_actix --features actix_extras,utoipa/auto_into_responses
 
     cargo test -p utoipa-gen --test path_derive_rocket --features rocket_extras
 
     cargo test -p utoipa-gen --test path_derive_axum_test --features axum_extras
-    cargo test -p utoipa-gen --test path_derive_auto_types_axum --features axum_extras,auto_types
+    cargo test -p utoipa-gen --test path_derive_auto_into_responses_axum --features axum_extras,utoipa/auto_into_responses
 elif [[ "$crate" == "utoipa-swagger-ui" ]]; then
     cargo test -p utoipa-swagger-ui --features actix-web,rocket,axum
 fi
