+++ utoipa-gen/tests/path_derive.rs
@@ -1,5 +1,3 @@
-#![cfg(not(feature = "auto_types"))]
-
 use std::collections::BTreeMap;
 
 use assert_json_diff::{assert_json_eq, assert_json_matches, CompareMode, Config, NumericMode};
