+++ utoipa-gen/tests/path_parameter_derive_actix.rs
@@ -1,5 +1,4 @@
 #![cfg(feature = "actix_extras")]
-#![cfg(not(feature = "auto_types"))]
 
 use utoipa::OpenApi;
